Added separate AutoMapper mapping package for each layer. Added DTO to logical object mapping in WebApi mapping package. Now WebApi references logical layer correctly.
